Combine Satellite runs into one group per Satellite
Previously we created one run for each host, regardless of whether it
was on a Satellite or not.  This may cause too many requests, and the
multiple jobs would be hard to track.  This now rearranges the run
list so that Satellite jobs are combined into one job per Satellite.

The tests now compare individual recipient data structures rather than
one big JSON list.

Signed-off-by: Paul Wayper <paulway@redhat.com>