RHINENG-17624: Better documentation on ack and hostack views and serialisers
Better documentation on ack and hostack views and serialisers

We also add the 'disabled_rules/' endpoint:

- GET /api/insights/v1/disabled_rules/ - a paginated list of disabled rules, by rule ID
- GET /api/insights/v1/disabled_rules/<rule_id>/ the details for a specific rule.

Each rule listed also gets a 'scope' field, which is either 'account' for
Acks (i.e. rule is disabled across all systems) or 'system' for Hostacks
(i.e. rule is disabled specifically on one or more systems).  If the detail
page is requested for a rule that is not disabled by either method, then
this gives a 404.

Signed-off-by: Paul Wayper <paulway@redhat.com>