Allow content import from an HTTP server
To support future updates of content via a content container, we now
support the 'import_content' command being supplied with a URL via the
`--remote` option.  This will then look for (in order):

1. `<url>config.yaml` - for the content configuration file.
2. `<url>rule_content.yaml.gz` or
3. `<url>rule_content.yaml` - for the actual rule content.
4. `<url>playbook_content.yaml.gz` or
5. `<url>playbook_content.yaml` - for the playbook content.

The rule and playbook content files will try the .gz compressed versions
first but will fail over.

The request will supply the `If-Modified-Since` header with the date of
the matching file on the filesystem, should it exist.  If a `304 Not
Modified` is received this is treated as a sign that we did not need to
download the content again.  However, the content and playbook import
will still go ahead - that code is relatively optimised.

The tests at the moment include basic checks that the data is imported
if the files do not exist, or exist and are unmodified.  More tests can
be added.

Signed-off-by: Paul Wayper <paulway@redhat.com>