The `import_content` code no longer runs as a script (#11)
Previously, the `api/scripts/import_content.py` script was expected to run by
itself and read files from the operating system. This is no longer supported.
However, the legacy of this was that we tried to manually set up the Django
environment, something that is now confusing parts of the WSGI loader. This
manual setup can easily be removed.

Signed-off-by: Paul Wayper <paulway@redhat.com>