Reduce the size of the logs in gunicorn, while increasing utility
For reasons as yet unknown, the logs from gunicorn include an 'args'
field that contains a dictionary with many many fields.  Some are
cryptic one-letter fields like 'U' and 'h'.  Others are named fields
in curly braces with a trailing letter, such as '{x-real-ip}i'.  Some
are duplicated in the headers.  Many we do not care about.

To preserve the same format of the logs that we currently store, we need
to copy these fields into the outer record as attributes with regular
names.

Signed-off-by: Paul Wayper <paulway@redhat.com>