Advisor fix logging conf hide metrics (#18)
The `hide_metrics` function seems to have been failing in Stage because it
assumed there were attributes of the log record which weren't there.

This fixes that problem and tests the full range of paths we expect through
this function.

Also fixes an incorrect string split attempt on the first character of a string,
and only rewrite URLs that are actually in the Tasks API

Signed-off-by: Paul Wayper <paulway@redhat.com>