Make sure the logs have the 'url' and 'method' fields. (#21)
For as yet unknown reasons, sometimes the gunicorn log arguments
have '-' as the value of the 'U' (URL) and 'm' (method) field. So we
fill those in from the 'raw_uri' and 'request_method' fields respectively.
We split the raw_uri into query parameters as well if it contains a '?'.

Signed-off-by: Paul Wayper <paulway@redhat.com>