Make sure fields we check/modify are in fields we might change (#25)
In investigating why the 'summary' field in many rules is the same as the
'generic' field, I found that changes to the 'summary' field were not being
tracked in trying to determine whether to update the rule model of an
existing rule. This fix makes sure that fields we check and possibly modify
are always included in the list of fields checked for updates.

(Hopefully) fixes RHINENG-17698

Signed-off-by: Paul Wayper <paulway@redhat.com>