Import Advisor content from remote URL path (#38)
This adds the ability of the import_content management command to request
rule config and content, and playbook content, from a remote URL base path.
The base URL is provided to the command using the `-u` or `--remote-url`
command line option.  The following URLs are tried (in order):

- `(BASE_URL)config.yaml`
- `(BASE_URL)rule_content.yaml.gz` or `(BASE_URL)rule_content.yaml` (in that order)
- `(BASE_URL)playbook_content.yaml.gz` or `(BASE_URL)playbook_content.yaml` (in that order)

At each stage, if all options on that line cannot be fetched (i.e. do not
return content with a status code of 200) then the content import fails (it
logs errors and exits).

The intent of this is to provide a content container that can serve the
latest rule and playbook content. This would allow us to then deprecate and
remove the current 'content server' system where a separate container POSTs
to the API via a hidden but unauthenticated URL.

We also add some tests of corner cases in the import_content command code to
improve our code test coverage.

Signed-off-by: Paul Wayper <paulway@redhat.com>