More logging to find cause of 'bad argument' fail (#67)
We're getting a message bad argument type for built-in operation in our logs
when we try to do an Org-level Kessel check. However, we don't see the log line
logger.info("KESSEL: checking %s has %s in %s", identity, permission, scope),
which makes me believe that for some reason the % operation implicit in
logger messages is somehow failing to stringify one of those three arguments.
We can see info level log messages elsewhere in the logs.

This now produces more info level messages, to make sure we're getting to
the has_kessel_permission function and to try and isolate which part of that
message might be causing problems.

Signed-off-by: Paul Wayper <paulway@redhat.com>