Kessel: Better errors, check environment specifically (#68)
In testing we're discovering that Stage had `KESSEL_ENABLED=False`, which
(now) causes the client to not start a connection (even if `KESSEL_URL` is set).
We need to give more explicit and specific error messages, as well as check the
`KESSEL_ENABLED` setting in the has_permission call checking whether the
feature flag is enabled.

It was possible to enable the feature flag in an environment which had not
initialised the Kessel client.

Signed-off-by: Paul Wayper <paulway@redhat.com>