Better authentication and permission failure messages
I discovered that the 'message' field on both the Authentication
and Permission classes in our main `permissions.py` code was (basically)
ignored by Django REST Framework.  The Authentication code should raise
AuthenticationError exceptions, and the Permission class did not have
the ability to raise a different message depending on different
circumstances (the 'message' attribute is read at class initialisation
time, not error reporting time).

We now report the permission check failure mode using the
'rbac_failure_message' property on the request, with this being added
to the log context.

We also remove the authentication classes from the Swagger views, since
we do not attempt to authenticate the user when providing the OpenAPI
schema.

Signed-off-by: Paul Wayper <paulway@redhat.com>