Validate UUID of RHC client ID from customer (#84)
The `rhc_id` field in the Sources message can be set by the user. We
see some invalid configurations where this is not a UUID, and this
causes a crash within the Tasks service. Guard against this by validating
the `rhc_id` field before trying to update our database.

Also detect an empty `source_ids` list and report that as well - there are
fewer instances of this in our error logs but it does happen occasionally.

Fixes RHINENG-21615

Signed-off-by: Paul Wayper <paulway@redhat.com>