Move topic editing to internal API (#83)
We recently got a report from RapiDAST detecting a 500 Server Error
when trying to POST a new topic as an unauthenticated user. This
revealed that we still basically allow anyone with an 'is_internal' flag to
create, update and delete Advisor rule topics, which is WAY too broad.

To solve RHINENG-20340 and also to reduce the number of people that
can edit topics, we now publish the editable API (POST, PATCH, PUT,
DELETE) only via the internal Turnpike API. The public API no longer
answers these methods and returns a 405 Method Not Allowed.

Signed-off-by: Paul Wayper <paulway@redhat.com>